Update README.md
This project seeks to determine some of the factors that effect how long a vehicle reamins listed for i.e. what causes a vehicle move off the market quickly or to stay listed for longer than usual. 
I took into account the price, the condition, number of cylinders and if the vehicle has 4wd. 
The data used can be found in the vehicles_us.csv file. I've also included the full jupyter notebook.